% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFreqTbl.R
\name{cleanFreqTbl}
\alias{cleanFreqTbl}
\title{clean a word frequency table to remove misspellings}
\usage{
cleanFreqTbl(freqTbl, outputFile = NULL, sdBoundary = 2)
}
\arguments{
\item{freqTbl}{frequency table}

\item{outputFile}{path for output, should be .R}

\item{sdBoundary}{integer for stringdist max distance to allow for frequencyTbl suggestions}
}
\description{
takes frequency table of words as input, performs
    interactive spellchecking function and outputs cleaned table.
    R script with changes during interaction is also generated.
}
