% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse2words.R
\name{df2words}
\alias{df2words}
\title{parse a df to words with specialized handling for scientific terms}
\usage{
df2words(df, ...)
}
\arguments{
\item{df}{dataframe object}

\item{deDupe}{boolean, default TRUE, dedupe df column elements prior to parsing}

\item{rmStopWords}{boolean, default TRUE, remove stopwords}
}
\description{
parse a df to words carefully to ensure that scientific terms
    which may include hyphens or plus-signs are not altered
}
