% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractiveFns.R
\name{InteractiveFindReplace.df}
\alias{InteractiveFindReplace.df}
\title{Interactively find and replace problematic words in a data frame}
\usage{
InteractiveFindReplace.df(misspelledWords, inputDF, outFile = NULL)
}
\arguments{
\item{misspelledWords}{named list of problematic words and suggested replacements}

\item{inputDF}{character vector}

\item{outFile}{filepath for where to append lines of code}
}
\description{
For each word in the output list from checkSpelling()
    find the word in the inputVector and allow the user to enter a
    replacement word
}
