% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractiveFns.R
\name{InteractiveFindReplace}
\alias{InteractiveFindReplace}
\title{Interactively find and replace problematic words in a vector or dataframe}
\usage{
InteractiveFindReplace(badWords, input, outFile = NULL)
}
\arguments{
\item{badWords}{named list of problematic words and suggested replacements}

\item{input}{dataframe or vector}

\item{outFile}{filepath for where to append lines of code}
}
\description{
For each word in the output list from checkSpelling()
    find the word in the inputVector and allow the user to enter a
    replacement word
}
