% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractiveFns.R
\name{interactiveSpellCheck.df}
\alias{interactiveSpellCheck.df}
\title{Interactively check spelling in dataframe against custom corpus}
\usage{
interactiveSpellCheck.df(input, outputDir, dfName = NULL)
}
\arguments{
\item{outputDir}{filepath for where to append lines of code}

\item{dfName}{name of df for use in output R doc, default NULL uses "templateName" attribute}

\item{inputDF}{dataframe, only character type will be worked on}
}
\description{
Given an input dataframe and output directory, the user may
    correct words that are not found in a standard dictionary or a
    current list of ImmuneSpace specific terms.
}
