% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractiveFns.R
\name{interactiveSpellCheck}
\alias{interactiveSpellCheck}
\title{Interactively check spelling against custom corpus}
\usage{
interactiveSpellCheck(input, name, outputDir, freqTbl)
}
\arguments{
\item{input}{dataframe or vector}

\item{name}{object name or string to use as fileName}

\item{outputDir}{filepath for where to append lines of code}

\item{freqTbl}{frequency table of words in a custom corpus}
}
\description{
Given an input vector or df and output directory, the user may
    correct words that are not found in a standard dictionary or a
    current list of ImmuneSpace specific terms.
}
