% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse2words.R
\name{list2words}
\alias{list2words}
\title{parse a list of character vectors to words with specialized handling for scientific terms}
\usage{
list2words(ls, ...)
}
\arguments{
\item{vec}{character vector}

\item{deDupe}{boolean, default TRUE, dedupe list elements prior to parsing}

\item{rmStopWords}{boolean, default TRUE, remove stopwords}
}
\description{
parse a list to words carefully to ensure that scientific terms
    which may include hyphens or plus-signs are not altered
}
