% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFreqTbl.R
\name{makeFreqTbl}
\alias{makeFreqTbl}
\title{make a frequency table of words from vector, list, or dataframe}
\usage{
makeFreqTbl(input, deDupe = FALSE, rmStopWords = FALSE)
}
\arguments{
\item{input}{R object of vector, list, or dataframe}

\item{deDupe}{boolean deduplicate}

\item{rmStopWords}{boolean remove stopwords}
}
\description{
takes an r-object of vector, list, or dataframe, and
    parses this into a frequency table of words
}
