% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse2words.R
\name{vec2words}
\alias{vec2words}
\title{parse a character vector to words with specialized handling for scientific terms}
\usage{
vec2words(vec, deDupe = TRUE, rmStopWords = TRUE)
}
\arguments{
\item{vec}{character vector}

\item{deDupe}{boolean, default TRUE, dedupe vector prior to parsing}

\item{rmStopWords}{boolean, default TRUE, remove stopwords}
}
\description{
parse a vector to words carefully to ensure that scientific terms
    which may include hyphens or plus-signs are not altered
}
