% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFreqTbl.R
\name{makeFreqTbl}
\alias{makeFreqTbl}
\title{make a frequency table of words from vector, list, or dataframe}
\usage{
makeFreqTbl(input, reduce = TRUE)
}
\arguments{
\item{input}{R object of vector, list, or dataframe}

\item{reduce}{boolean, default TRUE, controls de-duping of elements at highest level}
}
\description{
takes an r-object of vector, list, or dataframe, and
    parses this into a frequency table of words
}
